MLFlow_texture_bot/requirements.txt
# requirements.txt
torch
torchvision
Pillow
python-telegram-bot
nest-asyncio
numpy
mlflow
